==> B_nlpModel/3_cleanMetadata.ipynb <==
import re
import pandas as pd
import numpy as np

# Remove the special characters and images from a column. 
def removeSpecial(df_column):
    return df_column.astype(str).replace(r"[^\w\s,\.:!\/\&\-']+", "", regex=True).str.strip()

info_df = pd.read_csv('raw_info_df.csv').loc[:, ['jobTitle', 'description', 'employmentType', 'url', 'hiringOrganization.name', 'baseSalary.raw', 'baseSalary.valueMax', 'jobLocation.raw',]]

info_df.drop_duplicates(subset=['description'], inplace=True, ignore_index=True)

# These are the columns that can go through the removeSpecial function at the beginning.  
# The other columns are not yet ready - it's helpful to have the dollar symbol in front of a salary.
tidyColumnsBegin = ['description', 'employmentType', 'hiringOrganization.name', 'jobLocation.raw']

for col in tidyColumnsBegin:
    info_df[col] = removeSpecial(info_df[col])

# Updates for info_df['employmentType'] 
# Updated values to ensure value consistency. 
typeMask = info_df['employmentType'].str.contains("full", case=False)
info_df.loc[typeMask,['employmentType']] = "Full-Time"

# Removed values that do not align with the column title.
validEmploy = ['Remote', 'Full-Time', 'Freelance/Independent Contract']
info_df['employmentType'] = info_df['employmentType'].where(info_df['employmentType'].isin(validEmploy), np.nan)

# Splits the title to remove the location information. 
# The location information was checked against the df['jobLocation.raw'] rows with missing information, but no additional info was found.
info_df['jobTitle'] = info_df['jobTitle'].str.split('(', n=1, expand=True)[0]
info_df['jobTitle'] = removeSpecial(info_df['jobTitle'])

# Updated values for info_df['jobLocation.raw']
locMask = info_df['jobLocation.raw'].str.contains("Remote", case=False, na=None, regex=True)
info_df.loc[locMask,['jobLocation.raw']] = "Remote"

info_df['hiringOrganization.name'] = info_df['hiringOrganization.name'].str.replace("About", "", case=False, regex=True).str.strip()

# Extracting the salaries in the baseSalary column.
# "\$?\d{2,3},?\d{3}" For standard salary figures, an optional dollar sign and an optional comma, being 5-6 digits total
# "\$?\d{2,3}[kK]" For salaries expressed with k or K, digits with an optional dollar sign. 

info_df['baseSalary.raw'] = info_df['baseSalary.raw'].astype(str)
patternSalary = r"(\$?\d{2,3},?\d{3}|\$?\d{2,3}[kK])"

# Keep index to create ['level_0'], it is the index foreign index to info_df.
salary_df = info_df['baseSalary.raw'].str.extractall(patternSalary).reset_index(drop=False)

# Standardized extracted values, located in salary_df[0] 
salary_df[0] = salary_df[0].str.lower().replace({'401k': np.nan, 'k': '000', ',': ''}, regex=True)
salary_df.dropna(axis=0, inplace=True)

salary_df[0] = removeSpecial(salary_df[0])
salary_df[0] = salary_df[0].astype('int64')

minSalary = salary_df.groupby('level_0')[0].min().reset_index(name='minSalary')
maxSalary = salary_df.groupby('level_0')[0].max().reset_index(name='maxSalary')

salary_df = pd.merge(minSalary, maxSalary, on='level_0', how='outer').set_index('level_0')

# Joining the two dfs adds additional columns on the indexes of info_df and salary_df
df = info_df.join(salary_df).copy()
df.drop(['baseSalary.raw', 'baseSalary.valueMax'], axis=1, inplace=True)
df.columns = ['jobTitle', 'description', 'employmentType', 'url', 'hiringOrganization', 'jobLocation', 'minSalary', 'maxSalary']

# Reordered columns
df = df[['jobTitle', 'description', 'jobLocation', 'minSalary', 'maxSalary', 'hiringOrganization', 'employmentType', 'url', ]]
df.replace('nan', np.nan, inplace=True)

df['jobTitle'] = df['jobTitle'].str.replace(r"Sr\W", "Senior ", case=False, regex=True)\
                                .str.replace(r" - | & |/", " ", regex=True)\
                                .str.replace(r"[,]", "", regex=True)\
                                .str.replace(r"\s{2,}", " ", regex=True).str.strip()\
                                .replace("", "Data Scientist")# Adding the most frequent jobTitle for any missing jobTitles.

df.to_csv('metadataCleaned.csv', index=False)

df.info()

df.head(10)

df.nunique()
