==> B_nlpModel/4_deepClean.ipynb <==
import pandas as pd 
import spacy
import re

from spacy.pipeline import EntityRuler

nlp = spacy.load("en_core_web_md")

# Updated Spacy Entities
# Tokens like "Sklearn" and "IoT" were incorrectly labeled as spacy entities.
# These patterns update the token labels as "0" so they are not removed from the text.  
patterns = [{"label": "O", "pattern": "Sklearn"},
            {"label": "O", "pattern": "IoT"},
            {"label": "O", "pattern": "Databricks"},
            {"label": "O", "pattern": "Bioanalyzer"},
            {"label": "O", "pattern": "Biotech"},
            {"label": "O", "pattern": "BioHive"},
            {"label": "O", "pattern": "MLOps"},
            ]

ruler = nlp.add_pipe('entity_ruler', before='ner')
ruler.add_patterns(patterns)

#nlp.pipe_names

df = pd.read_csv('metadataCleaned.csv')

# Remove Sections of Text based on keyphrases. 

def removeEnts(text):
    doc = nlp(text)
    newText = [token.text + token.whitespace_ for token in doc if token.ent_type_ not in ["GPE", "LOC", "DATE", "CARDINAL", "ORDINAL"]]
    return "".join(newText)

def removeSection(text, phrase):
    pattern = rf'(?i)(\n.*?)({re.escape(phrase)}.*?)(\n|$)'
    newText = re.sub(pattern, '', text, flags=re.IGNORECASE)
    return newText

def removePhraseToSectionEnd(text, phrase):
    pattern = rf'(?i)({re.escape(phrase)}.*?)(\n|$)'
    newText = re.sub(pattern, '', text, flags=re.IGNORECASE)
    return newText


# Removed standard company statements on equal opportunity, benefit packages, and compensation to enhance data summary and model accuracy.
# Individually checked sections where the removed section exceeded the length of the remaining section.  Spot checked remaining sections.  
descriptionShort = df['description'].copy()

phrases = ["equal opportunity", "Equal Employment", "discriminate", "race, color, religion", "sexual orientation", "PTO ", "Benefits", "Compensation", "competitive", "COVID", "salary", "Health insurance", "applicants", "applicant", "dental", ] 

for p in phrases:
    descriptionShort = descriptionShort.apply(removeSection, phrase=p)
    descriptionShort = descriptionShort.apply(removePhraseToSectionEnd, phrase=p)

df['jobTitle'] = df['jobTitle'].apply(removeEnts)
df['descriptionShort'] = descriptionShort.apply(removeEnts)

# Text modifications
df['descriptionShort'] = df['descriptionShort'].str.replace(r"R&D", "Research and Development", case=False, regex=True)\
                                                .str.replace(r"&", "", regex=True)\
                                                .str.replace(r"A/B", "A-B", regex=True)\
                                                .str.replace(r"/", " ", regex=True)\
                                                .str.replace(r"\s{2,}", " ", regex=True).str.strip()\
                                                .str.replace(r"\n", " ", regex=True)

df.to_csv('dfCleaned.csv', index=False)

df.head()

# There are still 319 unique jobTitles for jobs that should have similar titles. 
df.nunique()

df.info()

# End DeepCleaning
# Finding Sections, Testing, and Notes

'''
def findEnts(text):
    doc = nlp(text)
    newText = [token.text for token in doc if token.ent_type_ in ["GPE", "LOC", "DATE", "CARDINAL", "ORDINAL"]]
    #if newText != '':
    return " ".join(newText)

def findSection(text, phrase):
    pattern = rf'(?i)(\n.*?)({re.escape(phrase)}.*?)(\n|$)'
    newText = re.findall(pattern, text, flags=re.IGNORECASE)
    newTextFormatted = ' '.join([''.join(match) for match in newText]) # Must do a double join to get the text as a string instead of a list
    return newTextFormatted

def findPhrase(text, phrase):
    pattern = rf'(?i)({re.escape(phrase)}.*?)(\n|$)'
    newText = re.findall(pattern, text, flags=re.IGNORECASE)
    newTextFormatted = ' '.join([''.join(match) for match in newText]) # Must do a double join to get the text as a string instead of a list
    return newTextFormatted

test = pd.DataFrame()
'''

#test['description find Ents'] = df['descriptionShort'].apply(findEnts)

#df['description'].apply(lambda x: re.findall("(\w+\/\w+)+", x, flags=re.IGNORECASE))#.value_counts()
#df['description'].apply(lambda x: re.findall("(\w+&\w+)+", x, flags=re.IGNORECASE)).value_counts()
#df['description'].apply(lambda x: re.findall("(\w+\-\w+)+", x, flags=re.IGNORECASE)).value_counts()

#test['find section']= df['descriptionShort'].apply(findSection, phrase='dental')

# These words were cleared. 
# 


#test.to_csv('test')

#test['find section'].value_counts()

#test['find PTO section'] = df['descriptionShort'].apply(findSection, phrase='PTO ')
#test['find Benefits section'] = df['descriptionShort'].apply(findSection, phrase='Benefits')
#test['find Compensation section'] = df['descriptionShort'].apply(findSection, phrase='Compensation')
